refactor(agents): extract promotion/finalization from cook.rs into cook_promotion
Second carve of the cook.rs god-file (after cook_budget). Lifts the entire
promote -> finalize boundary into a new cook_promotion module:
- promotion-source resolution (promotion_source/source_spec_path/source_worktree_path)
- the durable promote-or-load boundary (promote_attempt/promote_or_load_attempt/
  persisted_promotion_for_attempt)
- PR finalization (finalize_or_load_cook_pr*/finalize_cook_pr_with_backend)
- the cook_report builder + attempt_needs_execution

These sit downstream of a terminal provider result and publish controller-owned
state; grouping them keeps that boundary in one place.

Pure move, no behavior change: moved fns keep pub/pub(crate) visibility and are
re-exported through mod.rs's pub use, so crate::agent_task_service::* call sites
resolve unchanged. cook.rs imports back the handful it still calls
(cook_report, promote_or_load_attempt, finalize_or_load_cook_pr*, etc.);
cook.rs tests import the three they exercise. Trimmed the now-unused
finalization/promotion/review-dossier imports from cook.rs. cook.rs -350 lines
(3484 -> 3139); cook.rs is now 3139 vs the original 3575.

Verified: homeboy-agents --lib + --tests clean. The two cook promotion tests
(restarted_cook_uses_only_its_exact_persisted_promotion,
persisted_promotion_from_another_attempt_is_rejected) pass in isolation. The
agent_task_promotion git-op test failures are pre-existing (#8964 shared-state
flakes) and fail identically on the base branch without this change.
